move tmp file path check into persister factory
