+++ storage/factory/persisterFactory_test.go
@@ -2,8 +2,11 @@ package factory_test
 
 import (
 	"fmt"
+	"io/fs"
 	"os"
 	"path"
+	"path/filepath"
+	"strings"
 	"testing"
 
 	"github.com/multiversx/mx-chain-core-go/core/check"
@@ -36,6 +39,28 @@ func TestPersisterFactory_Create(t *testing.T) {
 		require.Equal(t, storage.ErrInvalidFilePath, err)
 	})
 
+	t.Run("with tmp file path, should work", func(t *testing.T) {
+		t.Parallel()
+
+		conf := createDefaultDBConfig()
+		conf.UseTmpAsFilePath = true
+
+		pf, _ := factory.NewPersisterFactory(conf)
+
+		dir := t.TempDir()
+
+		p, err := pf.Create(dir)
+		require.NotNil(t, p)
+		require.Nil(t, err)
+
+		// config.toml will be created in tmp path, but cannot be easily checked since
+		// the file path is not created deterministically
+
+		// should not find in the dir created initially.
+		_, err = os.Stat(dir + "/config.toml")
+		require.Error(t, err)
+	})
+
 	t.Run("should work", func(t *testing.T) {
 		t.Parallel()
 
@@ -46,7 +71,24 @@ func TestPersisterFactory_Create(t *testing.T) {
 		p, err := pf.Create(dir)
 		require.NotNil(t, p)
 		require.Nil(t, err)
+
+		// check config.toml file exists
+		_, err = os.Stat(dir + "/config.toml")
+		require.Nil(t, err)
 	})
+}
+
+func glob(root string) []string {
+	var files []string
+
+	filepath.WalkDir(root, func(s string, d fs.DirEntry, e error) error {
+		if filepath.Ext(s) == ".toml" {
+			files = append(files, s)
+		}
+		return nil
+	})
+
+	return files
 }
 
 func TestPersisterFactory_CreateWithRetries(t *testing.T) {
@@ -179,4 +221,26 @@ func TestPersisterFactory_IsInterfaceNil(t *testing.T) {
 
 	pf, _ := factory.NewPersisterFactory(createDefaultDBConfig())
 	require.False(t, pf.IsInterfaceNil())
+}
+
+func TestGetTmpFilePath(t *testing.T) {
+	t.Parallel()
+
+	pathSeparator := "/"
+
+	tmpDir := os.TempDir()
+	tmpBasePath := tmpDir + pathSeparator
+
+	path, err := factory.GetTmpFilePath("aaaa/bbbb/cccc")
+	require.Nil(t, err)
+	require.True(t, strings.Contains(path, tmpBasePath+"cccc"))
+
+	path, _ = factory.GetTmpFilePath("aaaa")
+	require.True(t, strings.Contains(path, tmpBasePath+"aaaa"))
+
+	path, _ = factory.GetTmpFilePath("")
+	require.True(t, strings.Contains(path, tmpBasePath+""))
+
+	path, _ = factory.GetTmpFilePath("/")
+	require.True(t, strings.Contains(path, tmpBasePath+""))
 }
